+++ bindings/java/lib/src/main/java/org/iota/types/exceptions/NoFundsReceivedFromFaucetException.java
@@ -1,0 +1,12 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+package org.iota.types.exceptions;
+
+public class NoFundsReceivedFromFaucetException extends Exception {
+
+    public NoFundsReceivedFromFaucetException() {
+        super();
+    }
+
+}