+++ src/message_interface/account_method.rs
@@ -405,4 +405,6 @@ pub enum AccountMethod {
     /// Expected response: [`SentTransaction`](crate::message_interface::Response::SentTransaction)
     #[cfg(feature = "participation")]
     DecreaseVotingPower { amount: String },
+    /// Expected response: [`Faucet`](crate::message_interface::Response::Faucet)
+    RequestFundsFromFaucet { url: String, address: String },
 }