+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `AccountBuilder::with_bech32_hrp()`;
 - `Account::retry_transaction_until_included()`;
 - `RetryTransactionUntilIncluded` to message interface account methods;
+- `AccountMethod::RequestFundsFromFaucet` to message interface;
 
 ### Changed